Add UserDTO classm/ JsonIdentityInfo
A new class, UserDTO, has been added to the DTOs package. This class will be used for transferring user data between different layers of the application.

The @JsonIdentityInfo annotation helps in handling circular references caused by bidirectional relationships in your entities. When you fetch a User, it includes the associated Authorities entity, which in turn includes all the User entities associated with it. This results in an infinite loop, causing the same User and Authorities entities to be repeated in your JSON output.

By using @JsonIdentityInfo, you’re instructing the JSON processor to output the ID of the entity instead of the entire entity once it has been output once already. This breaks the infinite loop and makes your JSON output more manageable and readable.